feat: 주문 결과 Alert 기능 구현 및 재고 차감 로직 오류 수정
